pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "lstm-multiple-layer-training"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.31.0",
  "pydantic==2.12.5",
  "PyYAML==6.0.3",
  "setuptools==80.10.2",
  "pytest==9.0.2",
  "pandas>=2.0.0",
  "scikit-learn>=1.3.0",
  "torch>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["program"]